refactor(tests): split the debate-log suite at the pure-function seam
Review finding on PR #5308: the suite reached 532 lines, over the
repository's 500-line file-size rule, so the claim that the taste ratchet
was unchanged did not hold for this tree. Measured before the split:

    [ERROR] test file-size: ...adr_debate_evidence.py:532
    File exceeds 500 lines (532 lines)

The split follows a real seam rather than an arbitrary line. Every case in
the new boundaries module exercises debate_log_evidence_gap and its helpers
as pure functions of text, with no git repository and no staged index; the
original keeps the cases that stage files and call check_adr_review_policy
end to end. 353 and 195 lines, taste-lints clean, 42 tests still passing.

The mutation harness now runs both halves. That is load-bearing rather than
bookkeeping: M6's killer, test_invalid_utf8_bytes_do_not_inflate_toward_the
_byte_floor, moved into the boundaries module, so a harness pointed at the
original alone would have reported M6 SURVIVED against working production
code. Its _assert_suite_ran guard caught the transient uncollected state
during this change, which is the guard doing its job.

Refs #5205
