+++ tests/validation/test_git_hook_policy_adr_debate_evidence.py
@@ -306,128 +306,6 @@ def test_adr_id_normalization(raw: str, expected: str) -> None:
     assert policy._normalize_adr_id(raw) == expected
 
 
-def test_the_weakest_currently_passing_log_is_pinned() -> None:
-    """Document the floor these four signals actually enforce.
-
-    This log is deliberately the thinnest thing that clears every signal: a
-    verdict label with a decision word inside the window, the token "agent" as
-    attribution, three headings, and filler to clear the byte floor. It says
-    nothing about any review that happened.
-
-    It passes, and that is the disclosed trade in ``debate_log_evidence_gap``:
-    every signal is a property of text the committer controls. The test exists
-    so the floor is visible and so a future tightening has a reference point to
-    move, rather than leaving the weakest passing shape undocumented.
-
-    The paired negative below pins the one thing that does constrain the shape:
-    the decision word has to sit within DEBATE_LOG_VERDICT_WINDOW_LINES of the
-    label, so a stray "Accepted" elsewhere in the file does not manufacture a
-    verdict.
-    """
-    weakest = (
-        "# D\n\n## Outcome\n\nAccepted.\n\n### Notes\n\nagent\n\n"
-        + "Filler prose about ADR-042 that carries no review content. " * 6
-    )
-    assert policy.debate_log_evidence_gap(weakest) is None
-
-
-def test_a_decision_word_outside_the_verdict_window_is_not_a_verdict() -> None:
-    """Negative control for the window that bounds the loosest passing shape."""
-    spread = (
-        "# D\n\n## Outcome\n\n### Notes\n\nagent\n\n"
-        + "\n" * 8
-        + "Accepted.\n"
-        + "Filler prose about ADR-042 that carries no review content. " * 6
-    )
-    gap = policy.debate_log_evidence_gap(spread)
-    assert gap is not None
-    assert "no verdict" in gap
-
-
-def _log_of_exactly(byte_count: int) -> str:
-    """Build a passing-shaped log padded to exactly ``byte_count`` bytes."""
-    head = "# A\n\n## Participants\n\narchitect\n\n### Verdict\n\nAccepted.\n"
-    pad = byte_count - len(head.encode("utf-8"))
-    assert pad >= 0, "head already exceeds the requested size"
-    return head + ("x" * pad)
-
-
-def test_the_byte_floor_is_exact() -> None:
-    """One byte under is rejected; exactly the floor is not."""
-    under = _log_of_exactly(policy.DEBATE_LOG_MIN_BYTES - 1)
-    at = _log_of_exactly(policy.DEBATE_LOG_MIN_BYTES)
-
-    assert len(under.encode("utf-8")) == policy.DEBATE_LOG_MIN_BYTES - 1
-    assert len(at.encode("utf-8")) == policy.DEBATE_LOG_MIN_BYTES
-
-    gap = policy.debate_log_evidence_gap(under)
-    assert gap is not None
-    assert "shorter than" in gap
-    assert policy.debate_log_evidence_gap(at) is None
-
-
-def test_the_section_floor_is_exact() -> None:
-    """Two headings are rejected; three are not, all else held equal."""
-    body = "\n\narchitect\n\nVerdict: Accepted.\n" + ("filler prose. " * 30)
-    two = "# A\n\n## Participants" + body
-    three = "# A\n\n## Participants\n\n### Round 1" + body
-
-    assert len(policy.DEBATE_LOG_HEADING_RE.findall(two)) == 2
-    assert len(policy.DEBATE_LOG_HEADING_RE.findall(three)) == 3
-
-    gap = policy.debate_log_evidence_gap(two)
-    assert gap is not None
-    assert "markdown sections" in gap
-    assert policy.debate_log_evidence_gap(three) is None
-
-
-@pytest.mark.parametrize(
-    ("offset", "expected_verdict"),
-    [
-        (policy.DEBATE_LOG_VERDICT_WINDOW_LINES - 1, True),
-        (policy.DEBATE_LOG_VERDICT_WINDOW_LINES, False),
-    ],
-)
-def test_the_verdict_window_is_exact(offset: int, expected_verdict: bool) -> None:
-    """The window is ``lines[i : i + N]``, so the last accepted offset is N-1.
-
-    Placing the decision one line further must not count, or the window is not
-    bounding anything.
-    """
-    lines = ["## Outcome"] + ["filler"] * (offset - 1) + ["Accepted."]
-    content = "\n".join(lines)
-    assert content.splitlines().index("Accepted.") == offset
-
-    assert policy._has_verdict(content) is expected_verdict
-
-
-@pytest.mark.parametrize(
-    "prose",
-    [
-        "The architecture is sound and the boundaries are well documented.",
-        "This securityless path needs no further hardening at all.",
-        "The analysis section explains the tradeoff in full detail here.",
-    ],
-)
-def test_prose_about_the_subject_is_not_reviewer_attribution(prose: str) -> None:
-    """A word that merely contains a role name must not stand in for a reviewer.
-
-    Without word boundaries, "architecture" satisfies "architect" and
-    "securityless" satisfies "security", so a log discussing the subject
-    matter would clear attribution while naming nobody.
-    """
-    assert not policy.DEBATE_LOG_REVIEWER_RE.search(prose)
-
-
-@pytest.mark.parametrize(
-    "attributed",
-    ["The architect reviewed it.", "Participants: two.", "Self-review: ACCEPT", "agent notes"],
-)
-def test_real_attribution_still_matches(attributed: str) -> None:
-    """Positive control: the boundaries must not break genuine attribution."""
-    assert policy.DEBATE_LOG_REVIEWER_RE.search(attributed)
-
-
 def test_an_incidental_mention_covers_a_staged_id(repo: Path) -> None:
     """Document the coverage rule's edge: any mention counts, reviewed or not.
 
@@ -447,63 +325,6 @@ def test_an_incidental_mention_covers_a_staged_id(repo: Path) -> None:
     )
 
     assert policy.check_adr_review_policy([ADR_42, ADR_05], repo) == 0
-
-
-def test_every_debate_log_in_the_working_tree_still_passes() -> None:
-    """Calibration pin: the thresholds must not false-block committed evidence.
-
-    Scope, stated precisely because the obvious reading is wrong: this reads
-    the corpus in the current working tree, not at any named ref. Local edits,
-    untracked logs, or a change that touches both the logs and this gate move
-    what it measures without main having moved. It is a guard against a
-    threshold change silently starting to reject real reviews, not a claim
-    about main's contents at any point in time.
-    """
-    critique = _ROOT / ".agents" / "critique"
-    logs = sorted(path for path in critique.glob("*.md") if "debate" in path.name)
-    assert len(logs) >= 70, "expected the calibration corpus to be present"
-
-    rejected = {
-        path.name: gap
-        for path in logs
-        if (gap := policy.debate_log_evidence_gap(path.read_text(errors="replace"))) is not None
-    }
-    assert rejected == {}
-
-
-def test_invalid_utf8_bytes_do_not_inflate_toward_the_byte_floor() -> None:
-    """A short blob of invalid bytes must not clear a floor it does not reach.
-
-    The staged blob is decoded with ``errors="replace"``, so each invalid byte
-    becomes U+FFFD and re-encodes to three. 100 on-disk bytes measured 300 and
-    cleared the stated 300-byte floor before ``_evidence_byte_count`` existed.
-    """
-    decoded = (b"\xff" * 100).decode("utf-8", errors="replace")
-    assert len(decoded.encode("utf-8")) == 3 * 100, "the inflation is what is being pinned"
-
-    gap = policy.debate_log_evidence_gap(decoded)
-    assert gap == f"shorter than {policy.DEBATE_LOG_MIN_BYTES} bytes"
-
-
-def test_replacement_characters_do_not_pad_a_real_log_over_the_floor() -> None:
-    """Negative pair: real text just under the floor stays under it when padded."""
-    body = "x" * (policy.DEBATE_LOG_MIN_BYTES - 1)
-    assert policy.debate_log_evidence_gap(body) is not None
-
-    padded = body + (b"\xff" * 50).decode("utf-8", errors="replace")
-    assert policy.debate_log_evidence_gap(padded) == (
-        f"shorter than {policy.DEBATE_LOG_MIN_BYTES} bytes"
-    )
-
-
-def test_valid_multibyte_text_still_counts_its_real_bytes() -> None:
-    """Positive control: the fix must not penalize genuine non-ASCII prose."""
-    # Well-formed U+FFFD written by an author counts as one character of real
-    # text, so stripping it is the one case where this measurement is stricter
-    # than the on-disk length. Everything else measures exactly.
-    text = "\u00e9" * 200
-    assert len(text.encode("utf-8")) == 400
-    assert policy._evidence_byte_count(text) == 400
 
 
 def test_frontmatter_only_implemented_flip_stays_exempt(repo: Path) -> None:
